+++ src/pages/Contact/index.jsx
@@ -1,19 +1,7 @@
 import { updateHead } from "../../utilities/updateHead";
-import React, { useEffect, useState } from "react";
-import { useForm } from "react-hook-form";
-import { yupResolver } from "@hookform/resolvers/yup";
-import * as yup from "yup";
-import { ContactSuccess, ContactForm, ContactLabel, ContactInput, ContactTextArea, ContactSubmit, ContactErrorReporting, ContactInputContainer, ContactReturnButton } from "./styles";
-
-// schema for form validation
-const schema = yup
-  .object({
-    fullName: yup.string().trim().min(3, "Full name is required and must be at least 3 characters.").required(),
-    Subject: yup.string().trim().min(3, "Subject is required and must be at least 3 characters.").required(),
-    Email: yup.string().trim().email().required(),
-    Body: yup.string().trim().min(3, "Message is required and must be at least 3 characters.").required(),
-  })
-  .required();
+import React, { useState } from "react";
+import { ContactSuccess, ContactFormContainer, ContactReturnButton } from "./styles";
+import ContactForm from "../../components/ContactForm/ContactForm";
 
 /**
  * Creates Contact page
@@ -21,15 +9,8 @@ const schema = yup
  */
 function ContactPage() {
   updateHead("Contact Us", "Contact TechBeauty with your queries, concerns and suggestions.");
+  // state to track if form is submitted successfully
   const [success, setSuccess] = useState(false);
-  const {
-    register,
-    handleSubmit,
-    reset,
-    formState: { errors, isSubmitSuccessful },
-  } = useForm({
-    resolver: yupResolver(schema),
-  });
 
   // submit form data
   function onSubmit(data) {
@@ -37,18 +18,11 @@ function ContactPage() {
     setSuccess(true);
   }
 
-  // reset form on successful submission
-  useEffect(() => {
-    if (isSubmitSuccessful) {
-      reset();
-    }
-  }, [isSubmitSuccessful, reset]);
-
   // if form is submitted successfully, show success message
   if (success) {
     return (
       <main>
-        <ContactForm onSubmit={handleSubmit(onSubmit)}>
+        <ContactFormContainer>
           <h1 id="main">Contact Us</h1>
           <ContactSuccess>Submitted Successfully</ContactSuccess>
           <p>
@@ -56,38 +30,18 @@ function ContactPage() {
             you.
           </p>
           <ContactReturnButton to="/">Return to store</ContactReturnButton>
-        </ContactForm>
+        </ContactFormContainer>
       </main>
     );
   }
 
-  // on load or if form is not submitted successfully, show form
+  // Render contact form
   return (
     <main>
-      <ContactForm onSubmit={handleSubmit(onSubmit)}>
+      <ContactFormContainer>
         <h1 id="main">Contact Us</h1>
-        <ContactInputContainer>
-          <ContactLabel htmlFor="full-name">Full Name</ContactLabel>
-          <ContactInput id="full-name" {...register("fullName")} />
-          <ContactErrorReporting>{errors.fullName?.message}</ContactErrorReporting>
-        </ContactInputContainer>
-        <ContactInputContainer>
-          <ContactLabel htmlFor="subject">Subject</ContactLabel>
-          <ContactInput id="subject" {...register("Subject")} />
-          <ContactErrorReporting>{errors.Subject?.message}</ContactErrorReporting>
-        </ContactInputContainer>
-        <ContactInputContainer>
-          <ContactLabel htmlFor="email">Email</ContactLabel>
-          <ContactInput id="email" {...register("Email")} />
-          <ContactErrorReporting>{errors.Email?.message}</ContactErrorReporting>
-        </ContactInputContainer>
-        <ContactInputContainer>
-          <ContactLabel htmlFor="body">Message</ContactLabel>
-          <ContactTextArea rows="5" id="body" {...register("Body")} />
-          <ContactErrorReporting>{errors.Body?.message}</ContactErrorReporting>
-        </ContactInputContainer>
-        <ContactSubmit type="submit">Submit</ContactSubmit>
-      </ContactForm>
+        <ContactForm onSubmit={onSubmit} />
+      </ContactFormContainer>
     </main>
   );
 }
